These tests don't test anything interesting.
They are running into issues because of a bug in moto but looking closer
at what they are testing.  Both tests test the happy path of the
underlying function.  In the happy path scenario all boto calls succeed
and all theses test do is test that but with all the boto calls mocked
out.  So we're just testing whether the mock is a good mock of boto
which is not useful.