:-(  Moto and HTTPretty don't play well together.
Moto has a vendored version of HTTPretty which also monkey patches the
socket object.  If you activate both, you get weird errors like
https://github.com/spulec/moto/issues/750

So we're opting not to activate both in this test.  We're are just using
the moto vendored version of httpretty for both our usage and the
mocking of boto.

I also removed the logic that seemed like it was just testing that our
mock is a mocking s3 correctly.